requirements.txt
# Data Pipeline
datasets>=2.19.0
huggingface-hub>=0.23.0
chromadb>=0.5.0
openai>=1.30.0
tiktoken>=0.7.0
requests>=2.31.0
tqdm>=4.66.0
pandas>=2.2.0

# Backend
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Agents
openai-agents>=0.1.0

# Document processing
pypdf>=4.2.0
python-docx>=1.1.0
markdown>=3.6

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
